Moved "node" and "react-native" into "src/platform"
