+++ packages/realm/src/platform/node/fs.ts
@@ -19,8 +19,8 @@
 import { existsSync, mkdirSync, readdirSync, rmSync, unlinkSync } from "node:fs";
 import { dirname, isAbsolute, join } from "node:path";
 
-import { inject } from "../platform/file-system";
-import { extendDebug } from "../debug";
+import { inject } from "../file-system";
+import { extendDebug } from "../../debug";
 
 const debug = extendDebug("fs");
