+++ packages/realm/src/platform/node/index.ts
@@ -15,9 +15,10 @@
 // limitations under the License.
 //
 ////////////////////////////////////////////////////////////////////////////
+
 import "./fs";
 import "./device-info";
 
-export * from "../index";
-import { Realm } from "../index";
+export * from "../../index";
+import { Realm } from "../../index";
 export default Realm;
