+++ packages/realm/src/platform/node/tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "extends": "../../tsconfig.json",
+  "extends": "../../../tsconfig.json",
   "compilerOptions": {
     "types": ["node"],
     "noResolve": false,
@@ -7,8 +7,8 @@
     "resolveJsonModule": true
   },
   "exclude": [
-    "../tests/",
+    "../../tests/",
     "../react-native/",
-    "../../type-tests/"
+    "../../../type-tests/"
   ],
 }
